add constant file and list options in descriptions
